disabled logging in verbose mode and made verbose mode the default mode
Verbose mode included info such as text rotation and angles but also printed/logged every extracted to stdout. This update disables the logging and adds a new run parameter '-log' to enable it. The "verbose" is also the default so text rotation and angle is enabled without having to specify it explicity